[MOVE/#13] MVVM 위한 리패키징
