+++ app/src/main/java/org/android/go/sopt/data/entity/local/mockPlayListTitle.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data.local
+package org.android.go.sopt.data.entity.local
 
 val mockPlayListTitle = listOf<PlaylistTitle>(
     PlaylistTitle(0, "상호의 플레이리스트")
