+++ app/src/main/java/org/android/go/sopt/data/service/AlbumService.kt
@@ -2,8 +2,8 @@ package org.android.go.sopt.data.service
 
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
-import org.android.go.sopt.data.remote.AddResponseDTO
-import org.android.go.sopt.data.remote.ListResponseDTO
+import org.android.go.sopt.data.entity.remote.AddResponseDTO
+import org.android.go.sopt.data.entity.remote.ListResponseDTO
 import retrofit2.http.*
 
 interface AlbumService {
