+++ app/src/main/java/org/android/go/sopt/data/service/AuthService.kt
@@ -1,9 +1,9 @@
 package org.android.go.sopt.data.service
 
-import org.android.go.sopt.data.remote.LoginRequestDTO
-import org.android.go.sopt.data.remote.LoginResponseDTO
-import org.android.go.sopt.data.remote.SignUpRequestDTO
-import org.android.go.sopt.data.remote.SignUpResponseDTO
+import org.android.go.sopt.data.entity.remote.LoginRequestDTO
+import org.android.go.sopt.data.entity.remote.LoginResponseDTO
+import org.android.go.sopt.data.entity.remote.SignUpRequestDTO
+import org.android.go.sopt.data.entity.remote.SignUpResponseDTO
 import retrofit2.http.Body
 import retrofit2.http.POST
