+++ app/src/main/java/org/android/go/sopt/presentation/follower/FollowerAdapter.kt
@@ -3,7 +3,7 @@ package org.android.go.sopt.presentation.follower
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import org.android.go.sopt.data.remote.FollowerResponseDTO
+import org.android.go.sopt.data.entity.remote.FollowerResponseDTO
 import org.android.go.sopt.databinding.ItemFollowerBinding
 import org.android.go.sopt.util.ItemDiffCallback
