+++ app/src/main/java/org/android/go/sopt/presentation/playlist/PlaylistAdapter.kt
@@ -3,7 +3,7 @@ package org.android.go.sopt.presentation.playlist
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import org.android.go.sopt.data.remote.MusicData
+import org.android.go.sopt.data.entity.remote.MusicData
 import org.android.go.sopt.databinding.ItemPlaylistSongBinding
 import org.android.go.sopt.util.ItemDiffCallback
